blockchain/main/rmgutil/wire: Use unsigned int to represent block height in the block header
Summary: Normally using an int32 to represent the block height is to represent an unknown height, however this is not required in the header as the height should always be known.

Reviewers: johann, ben

Reviewed By: ben

Differential Revision: https://phabricator.bitgo.com/D4260

+++ blockchain/difficulty.go
@@ -169,7 +169,7 @@ func (b *BlockChain) findPrevTestNetDifficulty(startNode *blockNode) (uint32, er
 	// Search backwards through the chain for the last block without
 	// the special rule applied.
 	iterNode := startNode
-	for iterNode != nil && iterNode.height%b.blocksPerRetarget != 0 &&
+	for iterNode != nil && int32(iterNode.height)%b.blocksPerRetarget != 0 &&
 		iterNode.bits == b.chainParams.PowLimitBits {
 
 		// Get the previous block node.  This function is used over
